+++ src/bitcoin/wallet-storage.ts
@@ -19,6 +19,18 @@ export function editWallet(currentAddress: string = '', addresses: number[] = []
   return parsedData;
 }
 
+export const updateStoredWallet = (key: string, value: string) => {
+  try {
+    const wallet = localStorage.getItem('wallet');
+    if (!wallet) return;
+    const parsedWallet = JSON.parse(wallet);
+    parsedWallet[key] = value;
+    localStorage.setItem('wallet', JSON.stringify(parsedWallet));
+  } catch (e) {
+    console.error(e);
+  }
+}
+
 export function saveWallet(mnemonic: string, network: string, currentAddress: string, addresses: number[], password: string) {
   const wallet = {
       mnemonic: encrypt(mnemonic, password),